+++ sockeye/arguments.py
@@ -766,7 +766,6 @@ def add_model_parameters(params):
 
     model_params.add_argument('--dtype', default=C.DTYPE_FP32, choices=[C.DTYPE_FP32, C.DTYPE_FP16],
                               help="Data type.")
-    # dtype-related safe clamp argument
     add_clamp_to_dtype_arg(model_params)
 
     model_params.add_argument('--amp',
@@ -1372,7 +1371,6 @@ def add_inference_args(params):
 
     decode_params.add_argument('--dtype', default=None, choices=[None, C.DTYPE_FP32, C.DTYPE_FP16, C.DTYPE_INT8],
                                help="Data type. Default: %(default)s infers from saved model.")
-    # dtype-related safe clamp argument
     add_clamp_to_dtype_arg(decode_params)
 
 
@@ -1412,9 +1410,9 @@ def add_brevity_penalty_args(params):
 def add_clamp_to_dtype_arg(params):
     params.add_argument('--clamp-to-dtype',
                         action='store_true',
-                        help='Clamp outputs for transformer attention, feed-forward, and process blocks to the min/max '
-                             'finite values for the current dtype. This can prevent inf/nan values from overflow when '
-                             'running large models in float16 mode. See: '
+                        help='Clamp outputs of transformer attention, feed-forward networks, and process blocks to the '
+                             'min/max finite values for the current dtype. This can prevent inf/nan values from '
+                             'overflow when running large models in float16 mode. See: '
                              'https://discuss.huggingface.co/t/t5-fp16-issue-is-fixed/3139')